+++ pkg/client/daemon/outbound.go
@@ -3,10 +3,12 @@ package daemon
 import (
 	"context"
 	"fmt"
+	"math/rand"
 	"net"
 	"strconv"
 	"strings"
 	"sync"
+	"time"
 
 	"github.com/pkg/errors"
 
@@ -155,7 +157,7 @@ func (o *outbound) firewallConfiguratorWorker(c context.Context) (err error) {
 			Routes: []*nat.Route{
 				{
 					Route: &rpc.Route{
-						Ip:    o.dnsIP,
+						Ips:   []string{o.dnsIP},
 						Proto: "udp",
 					},
 					Target: strconv.Itoa(o.dnsRedirPort),
@@ -198,6 +200,24 @@ func (o *outbound) resolveNoSearch(query string) *nat.Route {
 	return route
 }
 
+// Since headless and externalName services can have multiple IPs,
+// we return a shuffled list of the IPs if there are more than one.
+func (o *outbound) getIPs(ips []string) []string {
+	switch lenIPs := len(ips); lenIPs {
+	case 0:
+		return []string{}
+	case 1:
+	default:
+		// If there are multiple elements in the slice, we shuffle the
+		// order so it's not the same each time
+		rand.Seed(time.Now().UnixNano())
+		rand.Shuffle(lenIPs, func(i, j int) {
+			ips[i], ips[j] = ips[j], ips[i]
+		})
+	}
+	return ips
+}
+
 func (o *outbound) destination(conn *net.TCPConn) (string, error) {
 	_, host, err := o.translator.GetOriginalDst(conn)
 	return host, err
@@ -224,6 +244,50 @@ func (o *outbound) noMoreUpdates() {
 	close(o.work)
 }
 
+// Helper function for seeing if two unordered
+// slices of IPs are equal.
+func ipsEqual(a, b []string) bool {
+	// Since the slices could have duplicates, we need to
+	// create a slice of unique elements so we can compare them.
+	uniq := func(strSlice []string) []string {
+		uniqStrs := []string{}
+		uniqElms := make(map[string]bool, len(a))
+
+		for _, elm := range strSlice {
+			if _, ok := uniqElms[elm]; !ok {
+				uniqStrs = append(uniqStrs, elm)
+				uniqElms[elm] = true
+			}
+		}
+		return uniqStrs
+	}
+
+	uniqA := uniq(a)
+	uniqB := uniq(b)
+
+	if len(uniqA) != len(uniqB) {
+		return false
+	}
+
+	// Instead of comparing the slices directly, we make a
+	// map of the first slice, and then remove elements if
+	// they are in the second slice.
+	diff := make(map[string]bool, len(uniqA))
+	for _, aIP := range uniqA {
+		diff[aIP] = true
+	}
+	for _, bIP := range uniqB {
+		// ip was in uniqB but not uniqA, then they aren't equal
+		// and we can return early
+		if _, ok := diff[bIP]; !ok {
+			return false
+		}
+		delete(diff, bIP)
+	}
+	// If the diff map is empty at the end, then the ips are the same
+	return len(diff) == 0
+}
+
 func routesEqual(a, b *nat.Route) bool {
 	if a == b {
 		return true
@@ -231,7 +295,10 @@ func routesEqual(a, b *nat.Route) bool {
 	if a == nil || b == nil {
 		return false
 	}
-	return a.Name == b.Name && a.Ip == b.Ip && a.Port == b.Port && a.Target == b.Target
+	if !ipsEqual(a.Ips, b.Ips) {
+		return false
+	}
+	return a.Name == b.Name && a.Port == b.Port && a.Target == b.Target
 }
 
 func domain(r *nat.Route) string {
@@ -264,11 +331,11 @@ func (o *outbound) doUpdate(c context.Context, table *nat.Table) error {
 			if oldRouteOk {
 				switch newRoute.Proto {
 				case "tcp":
-					if err := o.translator.ClearTCP(c, oldRoute.Ip, oldRoute.Port); err != nil {
+					if err := o.translator.ClearTCP(c, oldRoute.Ips, oldRoute.Port); err != nil {
 						dlog.Errorf(c, "clear tpc: %v", err)
 					}
 				case "udp":
-					if err := o.translator.ClearUDP(c, oldRoute.Ip, oldRoute.Port); err != nil {
+					if err := o.translator.ClearUDP(c, oldRoute.Ips, oldRoute.Port); err != nil {
 						dlog.Errorf(c, "clear udp: %v", err)
 					}
 				default:
@@ -279,11 +346,11 @@ func (o *outbound) doUpdate(c context.Context, table *nat.Table) error {
 			if newRoute.Target != "" {
 				switch newRoute.Proto {
 				case "tcp":
-					if err := o.translator.ForwardTCP(c, newRoute.Ip, newRoute.Port, newRoute.Target); err != nil {
+					if err := o.translator.ForwardTCP(c, newRoute.Ips, newRoute.Port, newRoute.Target); err != nil {
 						dlog.Errorf(c, "forward tcp: %v", err)
 					}
 				case "udp":
-					if err := o.translator.ForwardUDP(c, newRoute.Ip, newRoute.Port, newRoute.Target); err != nil {
+					if err := o.translator.ForwardUDP(c, newRoute.Ips, newRoute.Port, newRoute.Target); err != nil {
 						dlog.Errorf(c, "forward udp: %v", err)
 					}
 				default:
@@ -314,11 +381,11 @@ func (o *outbound) doUpdate(c context.Context, table *nat.Table) error {
 
 		switch route.Proto {
 		case "tcp":
-			if err := o.translator.ClearTCP(c, route.Ip, route.Port); err != nil {
+			if err := o.translator.ClearTCP(c, route.Ips, route.Port); err != nil {
 				dlog.Errorf(c, "clear tpc: %v", err)
 			}
 		case "udp":
-			if err := o.translator.ClearUDP(c, route.Ip, route.Port); err != nil {
+			if err := o.translator.ClearUDP(c, route.Ips, route.Port); err != nil {
 				dlog.Errorf(c, "clear udp: %v", err)
 			}
 		default: