+++ pkg/client/daemon/resolved_linux.go
@@ -81,12 +81,12 @@ func (o *outbound) tryResolveD(c context.Context, onReady func()) error {
 
 	// DNS resolver
 	g.Go("Server", func(c context.Context) error {
-		v := dns.NewServer(c, []net.PacketConn{dnsResolverListener}, "", func(domain string) string {
+		v := dns.NewServer(c, []net.PacketConn{dnsResolverListener}, "", func(domain string) []string {
 			// Namespaces are defined on the network DNS config and managed by ResolveD, so not needed here.
 			if r := o.resolveNoSearch(domain); r != nil {
-				return r.Ip
+				return o.getIPs(r.Ips)
 			}
-			return ""
+			return []string{}
 		})
 		return v.Run(c)
 	})